This is a merge of multiple commits into one
Bellow are descriptions of commits that were squashed from oldest (on
top) to newest (on bottom)
---------------------------------------------------------------------

WIP - working on date-only/time-only

This version is working only if the input field has both date and time
in correct format, which is not what we want. The valid part should be
updated, the invalid discarded.

WIP - Not breaking if only valid date

This version is responsive even if only date is entered. Code needs to
be cleaned, this is only a middle step.

WIP - keyboard shortcuts, gui calendar, no clock

CTRL+C, CTRL+V, CTRL+A - standard
CTRL+SHIFT+C - toggle calendar
On copy creates a popup with "Copied to clipboard", disapears after 2s
Date works even without the clock, BUT for now only date will be in the
spinbox if some non-time values are inputed.

This is a working state, before I split the widget into multiple files
for more code clarity. Brace yourself - rollbacks incoming.

Popup changes

DImensions and position were adjusted and the popup apears behind the
main window, but only if the position extends to the main window. If
there is an empty space between main window and the popup, the popup
will be focus on