Tags included, preserve time
Tags now control what is available to display in a widget. We can change
date/time, date only and time only. Full precision tag controls if we
see miliseconds. Time is also preserved/updated when we change the date
in calendar. Calendar is only available if the `tag.date` is true.