Feature dateentry (#7)
datetime widget gui

* Tags included, preserve time

Tags now control what is available to display in a widget. We can change
date/time, date only and time only. Full precision tag controls if we
see miliseconds. Time is also preserved/updated when we change the date
in calendar. Calendar is only available if the `tag.date` is true.

* Removed width of entry fields

* Implementing method increment_part

This method is a result of refactoring change_date. Logic that checks if
the entry has a valid values is simplified and the amount of
code is reduced. If value is invalid, the entry won't change and no
exception is raised

* Fixed creating multiple entries

All widgets had another 2 invisible spinboxes. With time they were
visible right away, in case of date and datetime they were revealed with
opening a calendar

* Time changes hours by default

In spinbox with only time, default behaviour of the arrows is to
increase/decrease hours instead of minutes

---------

Co-authored-by: zf <zf-pentest@gmail.com>
Co-authored-by: zfydryn <zdenek.frydryn@nic.cz>