Fixed the select_all double insert
CTRL+V works like a usual pasting, additional inserting was causing to
paste and insert after the first one. The insert was there because
before it was possible to paste into a non-active entry, therefore the
content would have to be deleted and inserted. Now we only paste into a
in-focus entry.