Full precision rework
If the full precision tag is on, valid time is up to a six digits after
".", not necessarily exactly six.
When valid time is not found while clicking on calendar dates and we
want to have a valid time with or without precission, entry is updated
with zeros as time.