Feature dateentry (#8)
No font is specified now, so the all the spinbox field look like other
fields.
Working on login for showing/hiding time and time precision based on
tags

Removed run script from gitignore

Removed width of entry fields

Implementing method increment_part

This method is a result of refactoring change_date. Logic that checks if
the entry has a valid values is simplified and the amount of
code is reduced. If value is invalid, the entry won't change and no
exception is raised

Fixed creating multiple entries

All widgets had another 2 invisible spinboxes. With time they were
visible right away, in case of date and datetime they were revealed with
opening a calendar

Time changes hours by default

In spinbox with only time, default behaviour of the arrows is to
increase/decrease hours instead of minutes

Solved issue with multiple entries

When there are multiple DateEntryFrames on the page, shortcuts to toggle calendar (or copy) toggles the last one. This was achieved by passing a reference to a spinbox that will toggle the calendar.

Toggle calendar shortcut to active only

Calendar is now toggled with shortcut only on spinbox with focus

Remove of bind_all_events

All keyboard shortcuts now work only when the entry is in focus, which
is more intuitive solution.

Fixed the select_all double insert

CTRL+V works like a usual pasting, additional inserting was causing to
paste and insert after the first one. The insert was there because
before it was possible to paste into a non-active entry, therefore the
content would have to be deleted and inserted. Now we only paste into a
in-focus entry.

Keeping the caret position

When clicking on dates in calendar, caret position was always reseted
and moved to the end of the entry. This stores the caret positon and
move the caret there after the entry is updated.

Full precision rework

If the full precision tag is on, valid time is up to a six digits after
".", not necessarily exactly six.
When valid time is not found while clicking on calendar dates and we
want to have a valid time with or without precission, entry is updated
with zeros as time.

Full precision rework

If datetime/time has full precision, time is with seconds, otherwise
time is rounded to minutes without seconds.

Synchronizing files from origin

datetime gui tests