Feature dateentry (#8) (#9) active only
Time changes hours by default
Calendar is now toggled with shortcut only on spinbox with focus
All keyboard shortcuts now work only when the entry is in focus, which
is more intuitive solution.
Fixed the select_all double insert

CTRL+V works like a usual pasting, additional inserting was causing to
paste and insert after the first one. The insert was there because
before it was possible to paste into a non-active entry, therefore the
content would have to be deleted and inserted. Now we only paste into a
in-focus entry.

Keeping the caret position

When clicking on dates in calendar, caret position was always reseted
and moved to the end of the entry. This stores the caret positon and
move the caret there after the entry is updated.

Full precision rework

If the full precision tag is on, valid time is up to a six digits after
".", not necessarily exactly six.
When valid time is not found while clicking on calendar dates and we
want to have a valid time with or without precission, entry is updated
with zeros as time.

Co-authored-by: zfydryn <zdenek.frydryn@nic.cz>